added NC VC properties
