+++ iPhone Photos app/ForYouViewController.swift
@@ -14,5 +14,9 @@ class ForYouViewController: UIViewController {
         view.backgroundColor = .white
         title = "For You"
         navigationController?.navigationBar.prefersLargeTitles = true
+        navigationItem.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .add, target: self, action: #selector(libraryNCleftButtonTapped))
+        navigationItem.leftBarButtonItem?.tintColor = .black
     }
+    
+    @objc func libraryNCleftButtonTapped() { }
 }
