+++ iPhone Photos app/LibraryViewController.swift
@@ -14,5 +14,9 @@ class LibraryViewController: UIViewController {
         view.backgroundColor = .white
         title = "Library"
         navigationController?.navigationBar.prefersLargeTitles = true
+        navigationItem.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .add, target: self, action: #selector(libraryNCleftButtonTapped))
+        navigationItem.leftBarButtonItem?.tintColor = .black
     }
+    
+    @objc func libraryNCleftButtonTapped() { }
 }
