+++ iPhone Photos app/SearchViewController.swift
@@ -14,5 +14,9 @@ class SearchViewController: UIViewController {
         view.backgroundColor = .white
         title = "Search"
         navigationController?.navigationBar.prefersLargeTitles = true
+        navigationItem.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .add, target: self, action: #selector(libraryNCleftButtonTapped))
+        navigationItem.leftBarButtonItem?.tintColor = .black
     }
+    
+    @objc func libraryNCleftButtonTapped() { }
 }
